cuadro de dialogo para confirmar borrado del item

+++ frontend/src/app/components/items/items.component.ts
@@ -6,6 +6,8 @@ import { ItemService } from '../../../services/item.service';
 import { Router, ActivatedRoute, Params } from '@angular/router';
 import { Location } from '@angular/common';
 import { MatSnackBar } from '@angular/material';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-items',
@@ -27,6 +29,7 @@ export class ItemsComponent implements OnInit {
     private _route: ActivatedRoute,
     private location: Location,
     public snackBar: MatSnackBar,
+    public dialog: MatDialog
   ) {
     this.token = this._userService.getToken();
    }
@@ -105,19 +108,29 @@ export class ItemsComponent implements OnInit {
   }
 
   deleteItem(id,i){
-    this._itemService.deleteItem(id,this.token).subscribe(
-      response =>{
-        this.items.splice(i,1);
-        this.snackBar.open("Item borrado con exito.", '', {
-          duration: 500,
-        });          
-      },
-      error => {
-        this.snackBar.open(error.error.message, '', {
-          duration: 500,
-        });
+    let dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data:{message: '¿Desea borrar este item de la lista?'}
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        // ha pulsado SI, borramos
+        this._itemService.deleteItem(id,this.token).subscribe(
+          response =>{
+            this.items.splice(i,1);
+            this.snackBar.open("Item borrado con exito.", '', {
+              duration: 500,
+            });          
+          },
+          error => {
+            this.snackBar.open(error.error.message, '', {
+              duration: 500,
+            });
+          }
+        );
       }
-    );
+
+    });
+
   }
 
 }
